Remove Bad reasons section
This section is redundant as under Good reasons it mentions that compilers are dev tools and should not be on a system. The section is also very biased to the authors ideas and their own experiences and environments. Compilers not on systems for security is a very established practice and has corresponding policies in many environments and regulations such as DISA STIG's.